chore: split test file into multiple test files
