+++ packages/ucd-store/test/file-paths.test.ts
@@ -1,0 +1,104 @@
+import { setupMockStore } from "#internal/test-utils/store";
+import { UNICODE_VERSION_METADATA } from "@luxass/unicode-utils-new";
+import { UCDJS_API_BASE_URL } from "@ucdjs/env";
+import { createNodeUCDStore } from "@ucdjs/ucd-store";
+import { assert, beforeEach, describe, expect, it, vi } from "vitest";
+import { testdir } from "vitest-testdirs";
+
+describe("file paths", () => {
+  beforeEach(() => {
+    setupMockStore({
+      baseUrl: UCDJS_API_BASE_URL,
+      responses: {
+        "/api/v1/versions": [...UNICODE_VERSION_METADATA],
+      },
+    });
+
+    vi.clearAllMocks();
+    vi.unstubAllEnvs();
+  });
+
+  it("should get flattened file paths", async () => {
+    const storePath = await testdir({
+      "15.0.0": {
+        "ArabicShaping.txt": "Arabic shaping data",
+        "BidiBrackets.txt": "Bidi brackets data",
+        "extracted": {
+          "DerivedBidiClass.txt": "Derived bidi class data",
+          "nested": {
+            "DeepFile.txt": "Deep nested file",
+          },
+        },
+      },
+      ".ucd-store.json": JSON.stringify({
+        "15.0.0": "15.0.0",
+      }),
+    });
+
+    const store = await createNodeUCDStore({
+      basePath: storePath,
+    });
+
+    await store.init();
+    const [filePaths, filePathsError] = await store.getFilePaths("15.0.0");
+    assert(filePathsError === null, "Expected getFilePaths to succeed");
+    expect(filePaths).toEqual([
+      "ArabicShaping.txt",
+      "BidiBrackets.txt",
+      "extracted/DerivedBidiClass.txt",
+      "extracted/nested/DeepFile.txt",
+    ]);
+  });
+
+  it("should respect filters in getFilePaths", async () => {
+    const storePath = await testdir({
+      "15.0.0": {
+        "ArabicShaping.txt": "Arabic shaping data",
+        "BidiBrackets.txt": "Bidi brackets data",
+        "extracted": {
+          "DerivedBidiClass.txt": "Derived bidi class data",
+          "nested": {
+            "DeepFile.txt": "Deep nested file",
+          },
+        },
+      },
+      ".ucd-store.json": JSON.stringify({
+        "15.0.0": "15.0.0",
+      }),
+    });
+
+    const store = await createNodeUCDStore({
+      basePath: storePath,
+    });
+
+    await store.init();
+    const [filePaths, filePathsError] = await store.getFilePaths("15.0.0", ["!**/nested/**"]);
+    assert(filePathsError === null, "Expected getFilePaths to succeed");
+    expect(filePaths).toEqual([
+      "ArabicShaping.txt",
+      "BidiBrackets.txt",
+      "extracted/DerivedBidiClass.txt",
+    ]);
+  });
+
+  it("should throw error for invalid version in getFilePaths", async () => {
+    const storePath = await testdir({
+      "15.0.0": {
+        "ArabicShaping.txt": "Arabic shaping data",
+      },
+      ".ucd-store.json": JSON.stringify({
+        "15.0.0": "15.0.0",
+      }),
+    });
+
+    const store = await createNodeUCDStore({
+      basePath: storePath,
+    });
+
+    await store.init();
+    const [filePathsData, filePathsError] = await store.getFilePaths("16.0.0");
+    expect(filePathsData).toBe(null);
+    assert(filePathsError != null, "Expected error for invalid version");
+    expect(filePathsError.message).toBe("Version '16.0.0' does not exist in the store.");
+  });
+});
